alexi_rag/requirements.txt
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-groq
chromadb
python-dotenv
